+++ @ornikar/rollup-config/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [10.0.1](https://github.com/ornikar/shared-configs/compare/@ornikar/rollup-config@10.0.0...@ornikar/rollup-config@10.0.1) (2025-08-07)
+
+**Note:** Version bump only for package @ornikar/rollup-config
+
+
+
+
+
 ## [10.0.0](https://github.com/ornikar/shared-configs/compare/@ornikar/rollup-config@9.0.1...@ornikar/rollup-config@10.0.0) (2025-07-10)